+++ backend/users/Touir.ExpensesManager.Users/Controllers/Requests/ChangePasswordRequest.cs
@@ -5,7 +5,6 @@
         public string Email { get; set; }
         public string? OldPassword { get; set; }
         public string NewPassword { get; set; }
-        public string ConfirmPassword { get; set; }
 
     }
 }