+++ backend/users/Touir.ExpensesManager.Users/Controllers/Requests/ChangePasswordResetRequest.cs
@@ -5,7 +5,6 @@
         public string Email { get; set; }
         public string? VerificationHash { get; set; }
         public string NewPassword { get; set; }
-        public string ConfirmPassword { get; set; }
 
     }
 }