+++ backend/users/Touir.ExpensesManager.Users/Validators/RequestPasswordResetRequestValidator.cs
@@ -1,0 +1,14 @@
+using FluentValidation;
+using Touir.ExpensesManager.Users.Controllers.Requests;
+
+namespace Touir.ExpensesManager.Users.Validators
+{
+    public class RequestPasswordResetRequestValidator : AbstractValidator<RequestPasswordResetRequest>
+    {
+        public RequestPasswordResetRequestValidator()
+        {
+            RuleFor(x => x.Email).NotEmpty().WithMessage("MISSING_PARAMETERS");
+            RuleFor(x => x.AppCode).NotEmpty().WithMessage("MISSING_PARAMETERS");
+        }
+    }
+}